% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_climate_layers}
\alias{add_climate_layers}
\title{Add layers of climate data to GBIF occurrence data}
\usage{
add_climate_layers(climate_data, occurrence_data)
}
\arguments{
\item{climate_data}{a SpatRaster of climate layers}

\item{occurrence_data}{a tibble of GBIF occurrence data}
}
\value{
the original \code{occurrence_data} with an additional column for each
layer in \code{climate_data}
}
\description{
For all layers (i.e. climate variables) within \code{climate_data}, add each
layers' values in a given raster cell to all GBIF occurrence records with
coordinates contained by the cell.
}
\details{
GBIF records (i.e. rows) with missing values in \code{decimalLongitude} and
\code{decimalLatitude} are returned without the addition of climate layers.

\code{add_climate_layers()} is an internal function and is not intended to be
called directly by users.
}
\keyword{internal}
