% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-covariate-area.R
\name{calculate_covariate_area}
\alias{calculate_covariate_area}
\title{Calculate the area for each value of a covariate}
\usage{
calculate_covariate_area(
  covariate_raster = NULL,
  covariate = c("kgc", "pb", "phz")
)
}
\arguments{
\item{covariate_raster}{a terra raster object of covariate data}

\item{covariate}{a string representing the covariate, either 'kgc', 'pb', or
phz'}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
