% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_occurrence_data}
\alias{check_occurrence_data}
\title{Perform checks of the occurrence data}
\usage{
check_occurrence_data(occurrence_data)
}
\description{
Perform checks of the occurrence data
}
\keyword{internal}
