% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_layers.R
\name{classify_layers}
\alias{classify_layers}
\title{Classify each layer of climate data into suitability classes}
\usage{
classify_layers(climate_data, occurrence_data)
}
\arguments{
\item{climate_data}{a \code{SpatRaster} object containing climate data layers.}

\item{occurrence_data}{a tibble of GBIF occurrence records}
}
\value{
a list of tibbles, one per climate layer, each with level-wise
suitability classifications
}
\description{
Based on the proportion of a taxon's occurrence records within each level of
each layer of climate data, classify each level into one of four suitability
classifications: "Not suitable", "Not likely to be suitable",
"Likely to be suitable", and "Suitable".
}
\examples{
classify_layers(climate_layers_ssp370_2011_2040(), bryonia_alba)

}
