% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_levels.R
\name{classify_levels}
\alias{classify_levels}
\title{Classify the suitability of each level of climate data}
\usage{
classify_levels(climate_data, occurrence_data)
}
\arguments{
\item{climate_data}{a \code{SpatRaster} object containing climate data layers.}

\item{occurrence_data}{a tibble of GBIF occurrence records}
}
\value{
a list of tibbles, one per climate layer, each with level-wise
suitability classifications
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Based on the proportion of a taxon's occurrence records within each level of
each layer of climate data, classify each level into one of four suitability
classifications: "Not suitable", "Not likely to be suitable",
"Likely to be suitable", and "Suitable".
}
\details{
details go here
}
\examples{
classify_levels(climate_layers_ssp370_2011_2040(), bryonia_alba)

}
