% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_area}
\alias{compute_area}
\title{Compute the area covered by each level within each layer of climate data}
\usage{
compute_area(climate_data = NULL, unit = NULL)
}
\arguments{
\item{climate_data}{a terra SpatRaster containing one or more layers of
categorical climate data}

\item{unit}{optional, the abbreviation of the areal unit of measurement}
}
\value{
a list of tibbles, one per layer of climate data
}
\description{
For all layers (i.e. climate variables) within \code{climate_data}, calculate the
area of the Earth's surface covered by each level (i.e. category). By
default, area is calculated in squared meters, but any valid unit of
measurement in \code{units::valid_udunits()} can be supplied to the \code{unit}
argument.
}
\keyword{internal}
