% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_cells.R
\name{classify_cells}
\alias{classify_cells}
\title{Classify each cell of climate data into suitability classes}
\usage{
classify_cells(climate_data, layer_classes)
}
\arguments{
\item{climate_data}{a \code{SpatRaster} object containing climate data layers.}

\item{layer_classes}{a list of tibbles, one per climate layer, each with
level-wise suitability classifications}
}
\value{
a list of tibbles, one per climate layer, each with level-wise
suitability classifications
}
\description{
Using the suitability classifications in \code{layer_classes}, assign each cell of
climate \code{climate_data} a suitability classification for each layer of data.
Take the minimum suitability classification across all layers to assign an
overall classification to each cell.
}
\examples{
layer_classes <- classify_layers(
  climate_layers_ssp370_2011_2040(),
  bryonia_alba
)

classify_cells(climate_layers_ssp370_2011_2040(), layer_classes)

}
