% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_layers_ssp370_2011_2040.R
\name{climate_layers_ssp370_2011_2040}
\alias{climate_layers_ssp370_2011_2040}
\title{Climate layers averaged over 2011 - 2040 under the SSP370 emission scenario}
\source{
Derived from \url{https://zenodo.org/records/21400412}
}
\usage{
climate_layers_ssp370_2011_2040()
}
\value{
A \code{terra::SpatRaster} with three categorical layers, in
lon/lat WGS 84 (EPSG:4326).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Loads a categorical raster containing global Köppen-Geiger classes,
precipitation  bands, and plant hardiness zones. Downscaled and used for
illustrative purposes only. For higher resolution data, see
\url{https://zenodo.org/records/21400412}
}
\details{
A terra raster with three layers
\describe{
\item{\code{kgc_mode}}{Köppen-Geiger climate classes (1-30).}
\item{\code{pb_mode}}{Precipitation bands (0-11).}
\item{\code{phz_mode}}{Plant hardiness zone (1-28).}
}

details go here
}
\examples{
library(terra)

plot(climate_layers_ssp370_2011_2040()$pb_mode)

}
\concept{example data sets}
