% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protoclim_package.R
\docType{package}
\name{protoclim-package}
\alias{protoclim}
\alias{protoclim-package}
\title{protoclim: Climate Matching to Estimate Geographic Distributions of Plants}
\description{
Climate matching of a plant species using the three categorical climate variables adopted by the Proto3 model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cfia-pdam.github.io/protoclim/}
  \item \url{https://github.com/CFIA-PDAM/protoclim}
  \item Report bugs at \url{https://github.com/CFIA-PDAM/protoclim/issues}
}

}
\author{
\strong{Maintainer}: Josh Persi \email{josh.persi@inspection.gc.ca} (\href{https://orcid.org/0000-0002-2700-6483}{ORCID})

Authors:
\itemize{
  \item Josh Persi \email{josh.persi@inspection.gc.ca} (\href{https://orcid.org/0000-0002-2700-6483}{ORCID})
  \item Megan Abergel \email{megan.abergel@inspection.gc.ca} (\href{https://orcid.org/0000-0002-0681-1889}{ORCID})
}

Other contributors:
\itemize{
  \item Canadian Food Inspection Agency [copyright holder, funder]
}

}
\keyword{internal}
